Added upgrades
Added configurations to allow for a timer on when a player can reset the Trial Vault
Added commands to allow server operator to change the config options ingame
Added saving/loading for Trial Vaults and time stamping players that have looted the players
Trial Vaults no longer turn to stone block and back to reset. Added functionality to edit the NBT data to erase the players that have looted.