Fixed double reset bug - 1.0.15
Modified the loadfile method in TrialVault. I suspect that the loadfile might have been tampering with the ability to reset vault twice.
Changed the english locale time strings back from all upper case back to Letter case

Tested on 1.21.4 using RTag 1.5.10
Reloaded the server several times via command and console
Deleted the plugin and world and started afresh and readded the plugin to test.

Was not able to duplicate the double reset bug for the Trial Vaults.